// apps/web/src/lib/api/generated/context-program/context-program.ts
export type programsIndexResponse200 = {
  data: ProgramsIndex200;
  status: 200;
};

export type programsIndexResponse401 = {
  data: AuthenticationExceptionResponse;
  status: 401;
};

export type programsIndexResponseSuccess = programsIndexResponse200 & {
  headers: Headers;
};
export type programsIndexResponseError = programsIndexResponse401 & {
  headers: Headers;
};

export type programsIndexResponse =
  | programsIndexResponseSuccess
  | programsIndexResponseError;